Move test helper 'config' inside './client' to avoid importing external files
