+++ client/state/comments/test/actions.js
@@ -22,7 +22,7 @@ import {
 	changeCommentStatus,
 } from '../actions';
 import { NUMBER_OF_COMMENTS_PER_FETCH } from '../constants';
-import { setFeatureFlag } from 'test/helpers/config';
+import { setFeatureFlag } from 'test-helpers/config';
 
 const SITE_ID = 91750058;
 const POST_ID = 287;
